pyproject.toml
[tool.poetry]
name = "fast-api-guide"
version = "0.1.0"
description = ""
authors = ["tashfeenrao"]
license = "mit"
readme = "README.md"
packages = [{include = "fast_api_guide"}]

[tool.poetry.dependencies]
python = "3.10"
fastapi = {extras = ["all"], version = "^0.85.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
